+++ CHANGELOG.md
@@ -26,6 +26,7 @@ CHANGELOG
 - Added `last` action to move the cursor to the last match
     - The opposite action `top` is renamed to `first`, but `top` is still
       recognized as a synonym for backward compatibility
+- Added `preview-top` and `preview-bottom` actions
 - Extended support for alt key chords: alt with any case-sensitive single character
   ```sh
   fzf --bind alt-,:first,alt-.:last